Move init command to separate file
