+++ cmd/zen-go/cmd_init_test.go
